% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DownloadEOLpages.R
\name{DownloadEOLpages}
\alias{DownloadEOLpages}
\title{Download EOL Pages}
\usage{
DownloadEOLpages(pages, to.file = TRUE, MyKey = NULL, verbose = TRUE, ...)
}
\arguments{
\item{pages}{EOL page numbers to download}

\item{to.file}{Whether to download data to a file}

\item{MyKey}{An optional user identification key to identify yourself to EOL}

\item{verbose}{An optional print statement during download}

\item{...}{Curl options passed on to \code{\link[RCurl]{getURL}}}
}
\value{
Either an XML file(s) downloaded to working directory or as an R object saved in the
workspace.
}
\description{
Download EOL Pages
}
\details{
\code{DownloadEOLpages} will download EOL pages based on the EOL unique identifyer
number (EOL ID). Each taxon is associated with a unique identifier.  These numbers are used to
match EOL pages with hierarchy pages and keep track of taxonomic changes.

To generate an api key (MyKey), register with EOL and find it under your profile.
}
\examples{
\dontrun{
# Download taxa files to working directory in R
DownloadEOLpages(c(1,2,3), to.file=TRUE, MyKey)
}
}
\seealso{
\code{\link{DownloadHierarchy}}
}

