% Generated by roxygen2 (4.0.1): do not edit by hand
\name{FirstTwo}
\alias{FirstTwo}
\title{Get Binomial Nomenclature}
\usage{
FirstTwo(name)
}
\arguments{
\item{name}{Any taxonomic unit}
}
\value{
Returns a name
}
\description{
This function strips a taxon name to the first two words.  Many times, and EOL species name will
come in with additional data (such as the describer, subspecies, variety, etc).  This will prune
it down to just two names.
}
\examples{
\dontrun{
FirstTwo("Galanthus cilicicus Baker")
FirstTwo("Galanthus peshmenii A.P.Davis & C.D.Brickell")
FirstTwo("Galanthus")
}
}

