% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetHierID.R
\name{GetHierID}
\alias{GetHierID}
\title{Gets Hierarchy page ID}
\usage{
GetHierID(MyHier)
}
\arguments{
\item{MyHier}{A single filename for downloaded hierarchy page}
}
\value{
Returns a hierarchical concept ID.
}
\description{
This function uses the name of the file to return numerical ID.
}
\examples{
\dontrun{
# Works with a single page
GetHierID("hier51323249.xml")

# Or works with a list or vector of names
data(MyHiers)
GetHierID(MyHiers)
}
}
