% Generated by roxygen2 (4.0.1): do not edit by hand
\name{GetIUCNStat}
\alias{GetIUCNStat}
\title{Gather IUCN Status for EOL Pages}
\usage{
GetIUCNStat(MyEOLs)
}
\arguments{
\item{MyEOLs}{A vector of filenames or a list of XMLs for downloaded EOL pages}
}
\value{
Returns a data frame with taxon, eol ID, and IUCN status
}
\description{
This function gathers IUCN threat statuses from EOL pages if they are reported.
}
\examples{
\dontrun{
data(MyEOLs)
GetIUCNStat(MyEOLs)
GetIUCNStat(MyEOLs[3])
}
}
\seealso{
\code{\link{GetRichnessScores}} \code{\link{GetCommonNames}} \code{\link{GetReferences}}
\code{\link{DataObjectOverview}}
}

