\name{MyEOLs}
\alias{MyEOLs}
\alias{MyEOLs.rda}
\alias{MyHiers}
\alias{MyHiers.rda}
\docType{data}
\title{Example EOL and Hierarchy pages}
\description{
MyEOLs is an example of a subset of EOL files that were saved as an R object.  It is a list of six animal species (Camelus bactrianus, Camelus dromedarius, Hippopotamus amphibius, Rattus rattus, Rana cascadae, Bufo bufo), each item in the list is a separate species XML. The names of items in the list are the associated EOL numbers.  You can use these numbers to download the pages directly, or go to the EOL website and see the interactive page (for example, http://eol.org/pages/344581/overview). 

MyHiers is an example of the hierarchy pages that you can get from the EOL API, saved as an R object.  This download uses NCBI hierarchy information (see DownloadHierarchy help pages for an example). The taxa in MyHiers is the same as in MyEOLs.  Note that any identification numbers or even taxon names may differ between these files, since EOL and NCBI are independent.  
}
\format{XML, R}
\keyword{datasets}
