% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneFileHierarchy.R
\name{OneFileHierarchy}
\alias{OneFileHierarchy}
\title{Creates A Taxonomic Hierarchy for a single Taxon}
\usage{
OneFileHierarchy(MyHier)
}
\arguments{
\item{MyHier}{A single of filename for downloaded hierarchy page}
}
\value{
Returns a data frame where each row is a hierarchical taxonomic unit.
}
\description{
This function uses a provider hierarchy file to gather information about all taxonomic levels 
and their associated IDs.
}
\examples{
\dontrun{
data(MyHiers)
OneFileHierarchy(MyHiers[1])
}
}
