% Generated by roxygen2 (4.0.1): do not edit by hand
\name{PingAPI}
\alias{PingAPI}
\title{Test the working order of the EOL API}
\usage{
PingAPI(MyKey = NULL)
}
\arguments{
\item{MyKey}{An optional user identification key to identify yourself to EOL}
}
\value{
Returns a message with success or failure
}
\description{
This function will send out a ping to the EOL API to make sure it is in working order. The EOL
server will return either a message of success or failure.
}
\examples{
\dontrun{
PingAPI()
#PingAPI(MyKey)
}
}

