% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Reol-package.R
\docType{package}
\name{reol-package}
\alias{reol}
\alias{reol-package}
\title{R interface to the Encyclopedia of Life.}
\description{
R interface to the Encyclopedia of Life.
}
\details{
\tabular{ll}{
 Package: \tab Reol\cr
 Type: \tab Package\cr
 Version: \tab 1.21\cr
 Date: \tab 2013-9-9\cr
 License: \tab GPL (>= 2)\cr
}
}
\author{
Barb Banbury \email{bbanbury@utk.edu}

Brian O'Meara
}
\keyword{package}

