% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataforWeb.R
\name{CollectDataforWeb}
\alias{CollectDataforWeb}
\alias{CNCount}
\alias{DOCount}
\alias{providerCount}
\alias{DataProcessing}
\title{Collect Data from EOL Pages for Website}
\usage{
CollectDataforWeb(MyEOL, do.higher.taxonomy = FALSE)

CNCount(res)

DOCount(res)

providerCount(res)

DataProcessing(res, do.higher.taxonomy)
}
\arguments{
\item{MyEOL}{A single filename for downloaded EOL pages}

\item{do.higher.taxonomy}{(logical) on whether to download higher level taxonomy}

\item{res}{XML object}

\item{...}{Curl options passed on to \code{\link[RCurl]{getURL}}}
}
\value{
Appends EOL data to a table.
}
\description{
These functions will read and scrape content off the downloaded EOL pages. This is for building
the data table on eoldata.org.
}
\examples{
\dontrun{
data(MyEOLs)
CollectDataforWeb(MyEOLs[1])
}
}
