% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CollectDataforWeb}
\alias{CNCount}
\alias{CollectDataforWeb}
\alias{DOCount}
\alias{DataProcessing}
\alias{providerCount}
\title{Collect Data from EOL Pages for Website}
\usage{
CollectDataforWeb(MyEOL, do.higher.taxonomy = FALSE)

CNCount(res)

DOCount(res)

providerCount(res)

DataProcessing(res, do.higher.taxonomy)
}
\arguments{
\item{MyEOL}{A single filename for downloaded EOL pages}

\item{res}{XML object}

\item{do.higher.taxonomy}{(logical) on whether to download higher level taxonomy}
}
\value{
Appends EOL data to a table.
}
\description{
These functions will read and scrape content off the downloaded EOL pages. This is for building
the data table on eoldata.org.
}
\examples{
\dontrun{
data(MyEOLs)
CollectDataforWeb(MyEOLs[1])
}
}

