\name{Reol-package}
\alias{Reol-package}
\alias{Reol}
\alias{reol}
\docType{package}
\title{Reol}
\description{
	R interface to the Encyclopedia of Life
}
\details{
	\tabular{ll}{
	Package: \tab Reol\cr
	Type: \tab Package\cr
	Version: \tab 1.21\cr
	Date: \tab 2013-9-9\cr
	License: \tab GPL (>= 2)\cr}
}
\author{
Barb Banbury, Brian O'Meara
Maintainer: Barb Banbury <bbanbury@utk.edu>
}
